+++ hand-tracking-geometries/src/lib/constants.ts
@@ -2,36 +2,6 @@ export const ctxLineSize = 2.5;
 export const objectScaleTick = 0.025;
 export const pinchDistanceThreshold = 0.08;
 
-export type WebcamResponse = {
-	success: boolean;
-	error?: string | Error;
-};
-
-export type HandLandmark = {
-	visibility: unknown;
-	x: number;
-	y: number;
-	z: number;
-};
-
-export type FingerDistance = {
-	fingerTip: HandLandmark;
-	distanceToThumb: number;
-	distanceToBase: number;
-};
-
-export type MultiHandedness = {
-	displayName: string | undefined;
-	index: number;
-	label: "Left" | "Right";
-	score: number;
-};
-
-export type GestureResponse = {
-	gesture: string;
-	data?: unknown;
-};
-
 export const HandGestures = Object.freeze({
 	// pinched scale down or up
 	PINCHED: "pinched",
@@ -48,15 +18,6 @@ export const HandGestures = Object.freeze({
 	FINGER_UP_RIGHT: "finger up right",
 	// ....
 });
-
-export type HandGestureType = (typeof HandGestures)[keyof typeof HandGestures];
-
-export type TransformDirection = "down" | "up" | "left" | "right";
-export type TransformationType = "scale" | "rotation";
-export type TransformParams = {
-	transformDirection: TransformDirection;
-	transformation: TransformationType;
-};
 
 export const digits = [
 	// thumb
